# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_knowledge_dto.DocumentResponse:
    properties:
      created_at:
        type: string
      error:
        type: string
      filename:
        type: string
      id:
        type: string
      source_type:
        type: string
      status:
        type: string
      token_count:
        type: integer
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.AddDomainRequest:
    properties:
      domain:
        type: string
      is_primary:
        type: boolean
    required:
    - domain
    type: object
  github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.BrandingResponse:
    properties:
      company_id:
        type: string
      email_sender_name:
        type: string
      favicon_url:
        type: string
      logo_url:
        type: string
      primary_color:
        type: string
      secondary_color:
        type: string
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.DomainResponse:
    properties:
      domain:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      verified_at:
        type: string
    type: object
  github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.UpdateBrandingRequest:
    properties:
      email_sender_name:
        type: string
      favicon_url:
        type: string
      logo_url:
        type: string
      primary_color:
        type: string
      secondary_color:
        type: string
    type: object
  github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: SaaS multi-tenant white-label de atendimento automatizado e omnichannel
    (WhatsApp/Instagram) com IA + RAG.
  title: API
  version: "1.0"
paths:
...
  /agents/{id}/knowledge-bases/{kbId}:
    delete:
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Knowledge base ID
        in: path
        name: kbId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: agent or knowledge base not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlink knowledge base from agent
      tags:
      - knowledge
    post:
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Knowledge base ID
        in: path
        name: kbId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: agent or knowledge base not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Link knowledge base to agent
      tags:
      - knowledge
...
  /branding:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: branding not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get own branding
      tags:
      - tenant
    put:
      consumes:
      - application/json
      parameters:
      - description: Branding
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.UpdateBrandingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.BrandingResponse'
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update branding
      tags:
      - tenant
  /branding/host:
    get:
      parameters:
      - description: Hostname (defaults to request Host header)
        in: query
        name: host
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: branding not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      summary: Get branding by host
      tags:
      - tenant
...
  /domains:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.DomainResponse'
              type: array
            type: object
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List domains
      tags:
      - tenant
    post:
      consumes:
      - application/json
      parameters:
      - description: Domain
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.AddDomainRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_tenant_dto.DomainResponse'
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "409":
          description: domain already registered
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add domain
      tags:
      - tenant
...
  /knowledge-bases/{id}/documents:
    get:
      parameters:
      - description: Knowledge base ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_knowledge_dto.DocumentResponse'
              type: array
            type: object
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: knowledge base not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List documents
      tags:
      - knowledge
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Knowledge base ID
        in: path
        name: id
        required: true
        type: string
      - description: File (PDF, DOCX, TXT, etc.)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: invalid id or missing file
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: knowledge base not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload document (file)
      tags:
      - knowledge
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"